src/main/java/com/mycompany/poe/Task.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package com.mycompany.poe;
import java.util.ArrayList;
import java.util.List;
import javax.swing.JOptionPane;


/**
 *
 * @author RC_Student_lab
 */
public class Task {
    //declare variables
    private String taskName;
    private int taskNumber;
    private String taskDescription;
     private String developerDetails;
     private int taskDuration;
    private String taskID;
    private String taskStatus;
    
    //static list to keep track of all task durations to calculate totals hours
    static List<Integer> allTaskDurations = new ArrayList<>();
    
    //constructor to initialize task details
    public Task(String taskName, int taskNumber, String taskDescription, String developerDetails, int taskDuration, String taskStatus) {
        
        // Validate the task description
        if (!checkTaskDescription(taskDescription)) {
            JOptionPane.showMessageDialog(null, "Invalid task description. Task creation failed.", "Error", JOptionPane.ERROR_MESSAGE);
            return; // If invalid, do not proceed with task creation
        }
        
        this.taskName = taskName;
        this.taskNumber = taskNumber;
        this.taskDescription = taskDescription;
        this.developerDetails = developerDetails;
        this.taskDuration = taskDuration;
        this.taskStatus = taskStatus;
        this.taskID = createTaskID();
        allTaskDurations.add(taskDuration);
    
    }
   //getter methods to access attributes
   public String getTaskName(){
       return taskName;
   }
   
   public int getTaskNumber(){
       return taskNumber;
   }
   
   public String getTaskDescription(){
       return taskDescription;
   }
   
   public String getDeveloperDetails(){
       return developerDetails;
   }
   
   public int getTaskDuration(){
       return taskDuration;
   }
   
   public String getTaskID(){
       return taskID;
   }
  
    public String getTaskStatus() {
        return taskStatus;
    }
   
   //method to return a string with full task detaiils
   public String getTaskDetails(){
       return "Task Number: " + taskNumber + 
               "\nTask Name:" + taskName +
               "\nTask Description: " +
               taskDescription + "\nDeveloper: " + developerDetails +
               "\nDuration: " + taskDuration + "hours" + "\nTask ID: " +
               "\nStatus: " + taskStatus;
   }
   
   // Method to validate task description length
    public boolean checkTaskDescription(String taskDescription) {
        boolean isValid = taskDescription != null && !taskDescription.isEmpty() && taskDescription.length() <= 50;
        if (isValid) {
            JOptionPane.showMessageDialog(null, "Task successfully captured", "Success", JOptionPane.INFORMATION_MESSAGE);
        } else {
            JOptionPane.showMessageDialog(null, "Please enter a task description of less than 50 characters", "Error", JOptionPane.ERROR_MESSAGE);
        }
        return isValid;
    }

    // Method to create and return the TaskID
    public String createTaskID() {
        String taskPrefix = taskName.length() >= 2 ? taskName.substring(0, 2).toUpperCase() : taskName.toUpperCase();
        String developerSuffix = developerDetails.length() >= 3 ? developerDetails.substring(developerDetails.length() - 3).toUpperCase() : developerDetails.toUpperCase();
        return taskPrefix + ":" + taskNumber + ":" + developerSuffix;
    }

    /// Method to return the total combined hours of all entered tasks
    public static int returnTotalHours(List<Integer> taskDurations) {
        int totalHours = 0;
        for (int duration : taskDurations) {
            totalHours += duration;
        }
        return totalHours;
    }

    // Method to clear all task durations
    public static void clearTaskDurations() {
        allTaskDurations.clear();
    }

   // Method to display the total combined hours of all tasks using JOptionPane
    public static void displayTotalHours(List<Integer> taskDurations) {
        int totalHours = returnTotalHours(taskDurations);
        JOptionPane.showMessageDialog(null, "Total hours for all tasks: " + totalHours + " hours", "Total Hours", JOptionPane.INFORMATION_MESSAGE);
    }


    // Method to display the full task details using JOptionPane
    public void printTaskDetails() {
        JOptionPane.showMessageDialog(null, getTaskDetails(), "Task Details", JOptionPane.INFORMATION_MESSAGE);
    }
}